+++ src/main/java/com/example/demo/models/ItineraryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItineraryRepository extends JpaRepository<Itinerary, Long> {
+public interface ItineraryRepository extends JpaRepository<Itinerary, Integer> {
 }
